+++ pkg/expense/expense.go
@@ -1,9 +1,23 @@
 package expense
 
+import "database/sql"
+
 type Expense struct {
 	ID     int      `json:"id"`
 	Title  string   `json:"title"`
 	Amount float64  `json:"amount"`
 	Note   string   `json:"note"`
 	Tags   []string `json:"tags"`
+}
+
+type handler struct {
+	DB *sql.DB
+}
+
+type Error struct {
+	Message string `json:"message"`
+}
+
+func NewApplication(db *sql.DB) *handler {
+	return &handler{db}
 }
